requirements.txt
fastapi
uvicorn
gunicorn
torch
torchvision
pillow
numpy
open_clip_torch
